pyproject.toml
[project]
name = "cotizador-api"
version = "0.1.0"
description = "Cotizador API"
authors = [
    {name = "Francisco"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlmodel (>=0.0.22)",
    "pydantic-settings (>=2.0.0)",
    "alembic (>=1.13.0)",
    "bcrypt (>=4.0.0)",
    "python-jose[cryptography] (>=3.3.0)",
    "email-validator (>=2.0.0)",
    # "psycopg2-binary (>=2.9.9)",  # Driver para PostgreSQL
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
